Fixing merge conflict

+++ app/models/arm.rb
@@ -42,13 +42,13 @@ class Arm < ActiveRecord::Base
     return !name.nil? && name.length > 0
   end
 
-  # def valid_minimum_visit_count?
-  #   return !visit_count.nil? && visit_count >= minimum_visit_count
-  # end
+  def valid_minimum_visit_count?
+    return !visit_count.nil? && visit_count >= minimum_visit_count
+  end
 
-  # def valid_minimum_subject_count?
-  #   return !subject_count.nil? && subject_count >= minimum_subject_count
-  # end
+  def valid_minimum_subject_count?
+    return !subject_count.nil? && subject_count >= minimum_subject_count
+  end
 
   def create_line_items_visit line_item
     # if visit_count is nil then set it to 1
@@ -199,7 +199,13 @@ class Arm < ActiveRecord::Base
   def remove_visit position
     self.update_attribute(:visit_count, (self.visit_count - 1))
     visit_group = self.visit_groups.find_by_position(position)
-    return visit_group.destroy
+    unless visit_group.appointments.reject{|x| !x.completed_at?}.empty?
+      self.errors.add(:completed_appointment, "exists for this visit.")
+      return false
+    else
+      # self.update_attribute(:visit_count, (self.visit_count - 1))
+      return visit_group.destroy
+    end
   end
 
   def populate_subjects
